@@ -1,57 +1,49 @@
 {
-  "name": "26-tmdb",
-  "version": "0.2.0",
+  "name": "vite-tmdb",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "server": "nodemon api/server.js",
+    "build": "vite build",
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
+  },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.5",
-    "axios": "^1.4.0",
-    "bcrypt": "^5.1.0",
+    "axios": "^1.5.0",
+    "bcrypt": "^5.1.1",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "express-async-handler": "^1.2.0",
-    "jsonwebtoken": "^9.0.0",
-    "pg": "^8.10.0",
+    "js-cookie": "^3.0.5",
+    "jsonwebtoken": "^9.0.2",
+    "pg": "^8.11.3",
     "pg-hstore": "^2.3.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^8.0.5",
-    "react-router-dom": "^6.11.1",
-    "react-scripts": "^4.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.2",
+    "react-router-dom": "^6.16.0",
+    "react-scripts": "^5.0.1",
     "redux-logger": "^3.0.6",
-    "sequelize": "^6.31.1"
-  },
-  "scripts": {
-    "start": "concurrently 'npm run server' 'npm run client'",
-    "server": "nodemon api/server.js",
-    "client": "react-scripts start",
-    "build": "react-scripts build"
-  },
-  "proxy": "http://localhost:3001/",
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "sequelize": "^6.33.0"
   },
   "devDependencies": {
-    "bootstrap": "^5.2.3",
-    "concurrently": "^6.4.0",
-    "dotenv": "^16.0.3",
-    "nodemon": "^2.0.15",
-    "react-bootstrap": "^2.7.4",
-    "sweetalert2": "^11.7.5"
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@vitejs/plugin-react-swc": "^3.3.2",
+    "bootstrap": "^5.3.2",
+    "concurrently": "^8.2.1",
+    "dotenv": "^16.3.1",
+    "eslint": "^8.45.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "nodemon": "^3.0.1",
+    "react-bootstrap": "^2.8.0",
+    "sweetalert2": "^11.7.28",
+    "vite": "^4.4.5"
   }
 }
