@@ -1,5 +1,6 @@
-require("dotenv").config();
-const S = require("sequelize");
+import dotenv from 'dotenv';
+dotenv.config();
+import S from 'sequelize';
 const db_name = process.env.DB_NAME;
 const db_host = process.env.HOST;
 
@@ -9,4 +10,4 @@ const db = new S(db_name, null, null, {
   logging: false,
 });
 
-module.exports = db;
+export default db;
