@@ -1,13 +1,12 @@
-const jwt = require("jsonwebtoken");
+import jwt from 'jsonwebtoken';
 
 const SECRET = process.env.SECRET;
 
-const generateToken = (payload) => {
-  const token = jwt.sign({ user: payload }, SECRET, { expiresIn: "4d" });
+export const generateToken = (payload) => {
+  const token = jwt.sign({ user: payload }, SECRET, { expiresIn: '4d' });
   return token;
 };
 
-const validateToken = (token) => {
+export const validateToken = (token) => {
   return jwt.verify(token, SECRET);
 };
-module.exports = { generateToken, validateToken };
