@@ -1,20 +1,19 @@
-const asyncHandler = require("express-async-handler");
-const {
-  searchUser,
-  createUser,
-  generateUserCookie,
-  getAllUsers,
-} = require("../services/userService");
+import asyncHandler from 'express-async-handler';
 
-exports.create_new_user = asyncHandler(async (req, res, next) => {
+//Services
+import userServices from '../services/userService.js';
+import { generateToken } from '../config/token.js';
+
+export const create_new_user = asyncHandler(async (req, res, next) => {
   try {
     const { user_name } = req.body;
-    let user = await searchUser(user_name);
+    let user = await userServices.searchUser(user_name);
     if (user) {
-      res.status(400).send("user already exist");
+      res.status(400).send('user already exist');
     } else {
       let user_data = req.body;
-      let newUser = await createUser(user_data);
+      let newUser = await userServices.createUser(user_data);
+
       res.status(200).send(newUser);
     }
   } catch (error) {
@@ -22,10 +21,10 @@ exports.create_new_user = asyncHandler(async (req, res, next) => {
   }
 });
 
-exports.login_user = asyncHandler(async (req, res, next) => {
+export const login_user = asyncHandler(async (req, res, next) => {
   try {
     let { user_name, password } = req.body;
-    let user = await searchUser(user_name);
+    let user = await userServices.searchUser(user_name);
 
     let validatedPassword = await user.validatePassword(password);
 
@@ -35,25 +34,28 @@ exports.login_user = asyncHandler(async (req, res, next) => {
         user_name: user.user_name,
       };
 
-      let userCookie = await generateUserCookie(payload);
+      let userCookie = generateToken(payload);
 
-      res.cookie("token", userCookie);
+      res.cookie('authToken', userCookie);
 
       res.status(200).send(payload);
     } else {
-      res.send("wrong password or user_name");
+      res.send('wrong password or user_name');
     }
   } catch (error) {
     console.log(error);
   }
 });
 
-exports.logout_user = asyncHandler(async (req, res) => {
-  res.clearCookie("token");
-  res.sendStatus(204);
+export const logout_user = asyncHandler(async (req, res) => {
+  if (!req.user) {
+    return res.send('User is not logged in');
+  }
+
+  return res.clearCookie('authToken', { path: '/', domain: 'localhost' });
 });
 
-exports.get_all_users = asyncHandler(async (req, res) => {
-  let signedInUsers = await getAllUsers();
+export const get_all_users = asyncHandler(async (req, res) => {
+  let signedInUsers = await userServices.getAllUsers();
   res.status(200).send(signedInUsers);
 });
