@@ -1,15 +1,18 @@
-const express = require("express");
+import express from 'express';
 const router = express.Router();
-const {
+//Middlewares
+import validateUser from '../config/validateUser.js';
+//Controllers
+import {
   show_user_favorites,
   add_favorite,
   remove_from_favorite,
-} = require("../controllers/favorites_controllers");
+} from '../controllers/favorites_controllers.js';
 
-router.get("/:user_name", show_user_favorites);
+router.get('/:user_name', show_user_favorites);
 
-router.post("/add", add_favorite);
+router.post('/add', add_favorite);
 
-router.delete("/remove/:movie_name", remove_from_favorite);
+router.delete('/remove/:movie_name', remove_from_favorite);
 
-module.exports = router;
+export default router;
