@@ -1,20 +1,24 @@
-if (process.env.NODE_ENV !== "production") {
-  require("dotenv").config();
-}
-const PORT = process.env.SERVER_PORT;
-const express = require("express");
-const server = express();
-const db = require("./config/db/index");
-const router = require("./routes/index");
-const cookieParser = require("cookie-parser");
-const cors = require("cors");
+import dotenv from 'dotenv';
+import express from 'express';
+import db from './config/db/index.js';
+import router from './routes/index.js';
+import cookieParser from 'cookie-parser';
+import cors from 'cors';
 
-server.use(cookieParser());
+//Config
+dotenv.config();
+const PORT = process.env.SERVER_PORT;
+const server = express();
+
+//Middlewares
+server.use(cors({ origin: 'http://localhost:5173', credentials: true }));
 server.use(express.json());
-server.use("/api", router);
-server.use(cors());
+server.use(cookieParser());
+server.use('/api', router);
+
+//Db config
 db.sync({ force: false }).then(() => {
-  console.log("ESCUCHANDO A LA DB");
+  console.log('ESCUCHANDO A LA DB');
   server.listen(PORT, () => {
     console.log(`listening on port : ${PORT}`);
   });
