@@ -1,19 +1,19 @@
-const Swal = require("sweetalert2");
+import Swal from 'sweetalert2';
 
 export const SignedInMessage = () => {
   const Toast = Swal.mixin({
     toast: true,
-    position: "top",
+    position: 'top',
     showConfirmButton: false,
     timer: 3000,
     didOpen: (toast) => {
-      toast.addEventListener("mouseenter", Swal.stopTimer);
-      toast.addEventListener("mouseleave", Swal.resumeTimer);
+      toast.addEventListener('mouseenter', Swal.stopTimer);
+      toast.addEventListener('mouseleave', Swal.resumeTimer);
     },
   });
   const alert = Toast.fire({
-    icon: "success",
-    title: "Signed in successfully",
+    icon: 'success',
+    title: 'Signed in successfully',
   });
 
   return alert();
@@ -22,17 +22,17 @@ export const SignedInMessage = () => {
 export const FailedLogin = () => {
   const Toast = Swal.mixin({
     toast: true,
-    position: "top",
+    position: 'top',
     showConfirmButton: false,
     timer: 3500,
     didOpen: (toast) => {
-      toast.addEventListener("mouseenter", Swal.stopTimer);
-      toast.addEventListener("mouseleave", Swal.resumeTimer);
+      toast.addEventListener('mouseenter', Swal.stopTimer);
+      toast.addEventListener('mouseleave', Swal.resumeTimer);
     },
   });
   const alert = Toast.fire({
-    icon: "error",
-    title: "invalid password or username please try again",
+    icon: 'error',
+    title: 'invalid password or username please try again',
   });
 
   return alert();
@@ -41,17 +41,17 @@ export const FailedLogin = () => {
 export const EmailError = () => {
   const Toast = Swal.mixin({
     toast: true,
-    position: "top",
+    position: 'top',
     showConfirmButton: false,
     timer: 3500,
     didOpen: (toast) => {
-      toast.addEventListener("mouseenter", Swal.stopTimer);
-      toast.addEventListener("mouseleave", Swal.resumeTimer);
+      toast.addEventListener('mouseenter', Swal.stopTimer);
+      toast.addEventListener('mouseleave', Swal.resumeTimer);
     },
   });
   const alert = Toast.fire({
-    icon: "error",
-    title: "this email already exists",
+    icon: 'error',
+    title: 'this email already exists',
   });
 
   return alert();
@@ -60,17 +60,17 @@ export const EmailError = () => {
 export const AddToFavoriteMessage = () => {
   const Toast = Swal.mixin({
     toast: true,
-    position: "top",
+    position: 'top',
     showConfirmButton: false,
     timer: 3000,
     didOpen: (toast) => {
-      toast.addEventListener("mouseenter", Swal.stopTimer);
-      toast.addEventListener("mouseleave", Swal.resumeTimer);
+      toast.addEventListener('mouseenter', Swal.stopTimer);
+      toast.addEventListener('mouseleave', Swal.resumeTimer);
     },
   });
   const alert = Toast.fire({
-    icon: "success",
-    title: "Added to Favorites ",
+    icon: 'success',
+    title: 'Added to Favorites ',
   });
 
   return alert();
@@ -79,17 +79,17 @@ export const AddToFavoriteMessage = () => {
 export const DeleteFavoriteMessage = () => {
   const Toast = Swal.mixin({
     toast: true,
-    position: "top",
+    position: 'top',
     showConfirmButton: false,
     timer: 3000,
     didOpen: (toast) => {
-      toast.addEventListener("mouseenter", Swal.stopTimer);
-      toast.addEventListener("mouseleave", Swal.resumeTimer);
+      toast.addEventListener('mouseenter', Swal.stopTimer);
+      toast.addEventListener('mouseleave', Swal.resumeTimer);
     },
   });
   const alert = Toast.fire({
-    icon: "success",
-    title: "Deleted from Favorites ",
+    icon: 'success',
+    title: 'Deleted from Favorites ',
   });
 
   return alert();
