@@ -1,7 +1,7 @@
-const { useState } = require("react");
+import { useState } from 'react';
 
 const useInput = () => {
-  const [value, setValue] = useState("");
+  const [value, setValue] = useState('');
 
   const onChange = (event) => {
     setValue(event.target.value);
