@@ -1,8 +1,12 @@
-import { useState } from "react";
-import apiConfig from "../apiConfig";
-import { useNavigate } from "react-router-dom";
-import useInput from "../hooks/useInput";
-import axios from "axios";
+//Dependencies
+import React from 'react';
+import axios from 'axios';
+import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+//Hooks
+import useInput from '../hooks/useInput';
+//Config
+import apiConfig from '../apiConfig';
 
 const Tvdiscover = ({ element }) => {
   const [searched, setSearched] = useState([]);
@@ -14,7 +18,7 @@ const Tvdiscover = ({ element }) => {
   const handleSearch = async (e) => {
     try {
       e.preventDefault();
-      let query = search.value.replace(" ", "%20");
+      let query = search.value.replace(' ', '%20');
       let searcheShow = await axios.get(
         `${baseUrl}/search/tv?api_key=${apiKey}&language=en-US&page=1&query=${query}&include_adult=false`
       );
