@@ -1,11 +1,13 @@
-import apiConfig from "../apiConfig";
-import { useNavigate } from "react-router-dom";
+import React from 'react';
+import { useNavigate } from 'react-router-dom';
+//Config
+import apiConfig from '../apiConfig';
 
 const Populars = ({ title, poster_path, id }) => {
   const { w500Image } = apiConfig;
   const navigate = useNavigate();
 
-  const a = () => {
+  const goToMovieInfo = () => {
     navigate(`/movieinfo/${id}`);
   };
 
@@ -14,7 +16,7 @@ const Populars = ({ title, poster_path, id }) => {
       <div>
         <h1>{title}</h1>
         <img src={w500Image(poster_path)} alt="movie_image"></img>
-        <button onClick={a}>view more</button>
+        <button onClick={goToMovieInfo}>view more</button>
       </div>
     </>
   );
